requirements.txt
Flask
Flask-SQLAlchemy
Flask-Admin
Flask-Login
Flask-Migrate
Flask-Session
Flask-WTF
WTForms
email-validator
Flask-Mail
Flask-Caching
Flask-Limiter
PyMySQL
python-dotenv
Werkzeug==2.2.2
Flask-OAuthlib
docker
redis
rq
pdfkit
gunicorn